+++ packages/codegen.go/src/time.ts
@@ -7,6 +7,7 @@ import { values } from '@azure-tools/linq';
 import * as go from '../../codemodel.go/src/index.js';
 import { contentPreamble, getSerDeFormat, recursiveUnwrapMapSlice } from './helpers.js';
 import { ImportManager } from './imports.js';
+import { CodegenError } from './errors.js';
 
 // represents the generated content for an operation group
 export class Content {
@@ -46,7 +47,7 @@ export async function generateTimeHelpers(codeModel: go.CodeModel, packageName?:
         needsUnixTimeHelper = true;
         break;
       default:
-        throw new Error(`unhandled date-time format ${dateTimeFormat}`);
+        throw new CodegenError('InternalError', `unhandled date-time format ${dateTimeFormat}`);
     }
   };